# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Odace Data Pipeline API",
    "description": "\n## Odace Data Pipeline Platform\n\nA cloud-based data pipeline platform for managing data ingestion, transformation, and processing.\n\n### Features\n\n* **Multi-layer architecture**: Bronze (ingestion), Silver (transformation), Gold (aggregation)\n* **Delta Lake storage**: Efficient data lake with ACID transactions\n* **API-driven**: RESTful API for pipeline orchestration\n* **Authenticated**: Secure API key authentication\n* **Scalable**: Runs on Google Cloud Run\n\n### Authentication\n\nAll endpoints require an API key in the Authorization header:\n\n```\nAuthorization: Bearer sk_live_YOUR_API_KEY\n```\n\n### Available Pipelines\n\n**Bronze Layer (Data Ingestion)**:\n- accueillants, geo, logement, gares, lignes, zones_attraction\n\n**Silver Layer (Data Transformation)**:\n- accueillants, geo, logement, gares, lignes, zones_attraction\n\n### Getting Started\n\n1. Obtain an API key from your administrator\n2. Include the API key in all requests: `Authorization: Bearer sk_live_...`\n3. Use `/api/pipeline/list` to see available pipelines\n4. Run pipelines with `/api/bronze/{pipeline_name}` or `/api/silver/{pipeline_name}`\n\nFor detailed documentation, visit the project repository.\n    ",
    "contact": {
      "name": "Odace Data Team",
      "email": "data@odace.example.com"
    },
    "license": {
      "name": "Internal Use Only"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/pipeline/run": {
      "post": {
        "tags": [
          "pipeline"
        ],
        "summary": "Run Full Pipeline",
        "description": "Run the full data pipeline (bronze \u2192 silver).\n\nArgs:\n    request: Pipeline run configuration\n    user_id: User ID from API key authentication\n    \nReturns:\n    Summary of pipeline execution including job ID",
        "operationId": "run_full_pipeline_api_pipeline_run_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/FullPipelineRunRequest"
                  }
                ],
                "title": "Pipeline Request",
                "default": {
                  "bronze_only": false,
                  "silver_only": false,
                  "force": false
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/pipeline/status/{run_id}": {
      "get": {
        "tags": [
          "pipeline"
        ],
        "summary": "Get Pipeline Status",
        "description": "Get status of a pipeline run.\n\nArgs:\n    run_id: Pipeline run ID\n    api_key: API key for authentication\n    \nReturns:\n    Pipeline status details",
        "operationId": "get_pipeline_status_api_pipeline_status__run_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PipelineStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/pipeline/history": {
      "get": {
        "tags": [
          "pipeline"
        ],
        "summary": "Get Pipeline History",
        "description": "Get recent pipeline execution history.\n\nArgs:\n    limit: Maximum number of entries to return\n    api_key: API key for authentication\n    \nReturns:\n    List of recent pipeline executions",
        "operationId": "get_pipeline_history_api_pipeline_history_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/pipeline/list": {
      "get": {
        "tags": [
          "pipeline"
        ],
        "summary": "List Pipelines",
        "description": "List all available pipelines.\n\nArgs:\n    layer: Optional layer filter (bronze/silver/gold)\n    api_key: API key for authentication\n    \nReturns:\n    List of available pipelines",
        "operationId": "list_pipelines_api_pipeline_list_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "layer",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Layer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PipelineListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/jobs": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "List Jobs",
        "description": "List recent jobs with progress.\n\nArgs:\n    limit: Maximum number of jobs to return\n    api_key: API key for authentication\n    \nReturns:\n    List of jobs with progress information",
        "operationId": "list_jobs_api_jobs_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response List Jobs Api Jobs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs/{job_id}": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Get Job",
        "description": "Get job details including all tasks.\n\nArgs:\n    job_id: Job ID\n    api_key: API key for authentication\n    \nReturns:\n    Job details with task breakdown",
        "operationId": "get_job_api_jobs__job_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Get Job Api Jobs  Job Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FullPipelineRunRequest": {
        "properties": {
          "bronze_only": {
            "type": "boolean",
            "title": "Bronze Only",
            "description": "Run only bronze layer",
            "default": false
          },
          "silver_only": {
            "type": "boolean",
            "title": "Silver Only",
            "description": "Run only silver layer",
            "default": false
          },
          "force": {
            "type": "boolean",
            "title": "Force",
            "description": "Force reprocessing",
            "default": false
          }
        },
        "type": "object",
        "title": "FullPipelineRunRequest",
        "description": "Request to run the full pipeline.",
        "example": {
          "bronze_only": false,
          "force": false,
          "silver_only": false
        }
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PipelineInfo": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "layer": {
            "$ref": "#/components/schemas/PipelineLayer"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "dependencies": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Dependencies"
          }
        },
        "type": "object",
        "required": [
          "name",
          "layer"
        ],
        "title": "PipelineInfo",
        "description": "Information about a registered pipeline.",
        "example": {
          "dependencies": [],
          "description": "Ingest French commune geographic data",
          "layer": "bronze",
          "name": "geo"
        }
      },
      "PipelineLayer": {
        "type": "string",
        "enum": [
          "bronze",
          "silver",
          "gold"
        ],
        "title": "PipelineLayer",
        "description": "Pipeline layer types."
      },
      "PipelineListResponse": {
        "properties": {
          "pipelines": {
            "items": {
              "$ref": "#/components/schemas/PipelineInfo"
            },
            "type": "array",
            "title": "Pipelines"
          }
        },
        "type": "object",
        "required": [
          "pipelines"
        ],
        "title": "PipelineListResponse",
        "description": "List of available pipelines."
      },
...
      "PipelineStatus": {
        "type": "string",
        "enum": [
          "pending",
          "running",
          "success",
          "failed"
        ],
        "title": "PipelineStatus",
        "description": "Pipeline execution status."
      },
      "PipelineStatusResponse": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "pipeline_name": {
            "type": "string",
            "title": "Pipeline Name"
          },
          "layer": {
            "$ref": "#/components/schemas/PipelineLayer"
          },
          "status": {
            "$ref": "#/components/schemas/PipelineStatus"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "duration_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Seconds"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": ""
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "stats": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stats"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "pipeline_name",
          "layer",
          "status",
          "started_at"
        ],
        "title": "PipelineStatusResponse",
        "description": "Response for pipeline status query.",
        "example": {
          "completed_at": "2025-11-29T10:35:00Z",
          "duration_seconds": 300.5,
          "layer": "bronze",
          "message": "Successfully processed 50 files",
          "pipeline_name": "geo",
          "run_id": "550e8400-e29b-41d4-a716-446655440000",
          "started_at": "2025-11-29T10:30:00Z",
          "stats": {
            "files_processed": 50,
            "files_skipped": 2,
            "records_ingested": 15000
          },
          "status": "success"
        }
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}